# docs/swagger.yaml
basePath: /
definitions:
  domain.Ingredient:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  domain.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        description: Stored separately in MongoDB, populated in memory
        items:
          $ref: '#/definitions/domain.OrderProduct'
        type: array
      phone_number:
        description: Stored alongside user_id as per diagram
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.OrderProduct:
    properties:
      adicionais:
        description: array of extra ingredient IDs
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      ingredients:
        description: array of ingredient IDs included
        items:
          type: string
        type: array
      name:
        type: string
      order_id:
        type: string
      price:
        description: calculated price (base product + adicionais)
        type: number
      product_id:
        type: string
      removed:
        description: array of removed ingredient IDs
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
  http.createIngredientReq:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  http.createOrderReq:
    properties:
      items:
        items:
          $ref: '#/definitions/ports.OrderItemInput'
        type: array
      user_id:
        type: string
    type: object
...
  ports.OrderItemInput:
    properties:
      adicionais:
        items:
          type: string
        type: array
      product_id:
        type: string
      removed:
        items:
          type: string
        type: array
    type: object
host: localhost:7000
info:
  contact: {}
  description: REST API for a restaurant backend
  title: Restaurant API
  version: "1.0"
paths:
  /api/v1/ingredients:
    get:
      description: Get a list of all ingredients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Ingredient'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all ingredients
      tags:
      - ingredients
    post:
      consumes:
      - application/json
      description: Create a new ingredient with name and price
      parameters:
      - description: Ingredient details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createIngredientReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Ingredient'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new ingredient
      tags:
      - ingredients
  /api/v1/ingredients/{id}:
    delete:
      description: Delete an ingredient by ID
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an ingredient
      tags:
      - ingredients
    get:
      description: Get details of a single ingredient by ID
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ingredient'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ingredient by ID
      tags:
      - ingredients
    put:
      consumes:
      - application/json
      description: Update ingredient details by ID
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: string
      - description: Ingredient details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createIngredientReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ingredient'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an ingredient
      tags:
      - ingredients
  /api/v1/orders:
    get:
      description: Get a list of all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with user ID and items
      parameters:
      - description: Order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createOrderReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new order
      tags:
      - orders
  /api/v1/orders/{id}:
    get:
      description: Get details of a single order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get order by ID
      tags:
      - orders
...
swagger: "2.0"